+++ test/fixtures.js
@@ -1,0 +1,55 @@
+module.exports = [
+    {
+        "title": "Soupe de courgette au kiri",
+        "description": "Une recette rapide de soupe de courgettes",
+        "language": "FR_fr",
+        "ingredients": [
+            {
+                "ingredient": "courgette",
+                "quantity": "3",
+                "unit": ""
+            },
+            {
+                "ingredient": "pomme de terre",
+                "quantity": "2",
+                "unit": ""
+            },
+            {
+                "ingredient": "fromage type kiri",
+                "quantity": "3",
+                "unit": ""
+            },
+            {
+                "ingredient": "oignon",
+                "quantity": "1",
+                "unit": ""
+            }
+        ],
+        "steps": [
+            {
+                "index": 0,
+                "description": "Découper les courgettes grossièrement"
+            },
+            {
+                "index": 1,
+                "description": "Peler et découper grossièrement l'oignon"
+            },
+            {
+                "index": 2,
+                "description": "Mettre les légumes dans la cocotte minute"
+            },
+            {
+                "index": 3,
+                "description": "Ajouter 3 fromage type kiri"
+            },
+            {
+                "index": 4,
+                "description": "Ajouter deux bouillons cubes"
+            },
+            {
+                "index": 5,
+                "description": "Recouvrir d'eau et laisser cuire 25 min dès que ca siffle"
+            }
+        ]
+    }
+]