+++ changelog.md
@@ -1,3 +1,4 @@
 # 0.1.0
 * create the project
-* can get the list of recipes
+* We can get the list of recipes
+* We can create a recipe